deployment/requirements.txt
torch
torchvision
streamlit
opencv-python
pillow
numpy
grad-cam
